split up goto_check
goto_check inserts assertions that check for suspect bugs in both C/C++ and
Java programs.  The criteria differ vastly given that Java has basically no
instances of undefined behavior.

This commit splits the file, separating it into a checker for C/C++ and one
for Java.  The long-term plan is to move these checkers into the directory
of the corresponding language frontend.
