+++ jbmc/src/jbmc/jbmc_parse_options.cpp
@@ -228,8 +228,8 @@ void jbmc_parse_optionst::get_command_line_options(optionst &options)
     "self-loops-to-assumptions",
     !cmdline.isset("no-self-loops-to-assumptions"));
 
-  // all checks supported by goto_check
-  PARSE_OPTIONS_GOTO_CHECK(cmdline, options);
+  // all checks supported by goto_check_java
+  PARSE_OPTIONS_GOTO_CHECK_JAVA(cmdline, options);
 
   // unwind loops in java enum static initialization
   if(cmdline.isset("java-unwind-enum-static"))
@@ -811,7 +811,7 @@ void jbmc_parse_optionst::process_goto_function(
   }
 
   // add generic checks
-  goto_check(
+  goto_check_java(
     function.get_function_id(),
     function.get_goto_function(),
     ns,
