+++ jbmc/src/jbmc/jbmc_parse_options.h
@@ -19,7 +19,7 @@ Author: Daniel Kroening, kroening@kroening.com
 
 #include <langapi/language.h>
 
-#include <analyses/goto_check.h>
+#include <analyses/goto_check_java.h>
 
 #include <goto-checker/bmc_util.h>
