+++ src/analyses/Makefile
@@ -13,6 +13,8 @@ SRC = ai.cpp \
       flow_insensitive_analysis.cpp \
       global_may_alias.cpp \
       goto_check.cpp \
+      goto_check_c.cpp \
+      goto_check_java.cpp \
       goto_rw.cpp \
       guard_bdd.cpp \
       guard_expr.cpp \
